fix: refactor the middlewares
