+++ backend/index.js
@@ -1,30 +1,16 @@
-const express = require('express');
-const bodyParser = require('body-parser');
-const app = express();
-const adminRouter = require("./routes/admin")
-const userRouter = require("./routes/user");
-const projectRouter = require("./routes/project");
-const taskRouter = require("./routes/task");
-const commentRouter = require("./routes/comment");
-const cors = require("cors");
-const { connectDB } = require('./db');
+// index.js
+const app = require('./app');
+const {connectDB} = require("../backend/db/index.js")
 
-// Middleware for parsing request bodies
-app.use(bodyParser.json());
-app.use(cors());
-app.use("/admin", adminRouter)
-app.use("/user", userRouter)
-app.use("/project", projectRouter)
-app.use("/task", taskRouter)
-app.use("/comment", commentRouter)
+// Port configuration
+const PORT = process.env.PORT || 3001;
 
-const PORT = 3001;
-
-// Start server only after connecting to MongoDB
 const startBackend = async () => {
     try {
-        await connectDB(); // Wait for database connection and try again and again
+        // Ensure the database is connected before starting the server
+        await connectDB();
 
+        // Start listening on the specified port
         app.listen(PORT, () => {
             console.log(`Server is running on port ${PORT}`);
         });
@@ -34,4 +20,5 @@ const startBackend = async () => {
     }
 };
 
+// Start the backend
 startBackend();
